+++ src/installer/detector.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { exec } from '../utils/utils';
+import { exec } from '../utils/helpers';
 
 export async function getLsVersion(binPath: string): Promise<string | undefined> {
   try {
