+++ src/utils/vscode.ts
@@ -4,18 +4,6 @@ export function config(section: string, scope?: vscode.ConfigurationScope): vsco
   return vscode.workspace.getConfiguration(section, scope);
 }
 
-export function getFolderName(folder: vscode.WorkspaceFolder): string {
-  return normalizeFolderName(folder.uri.toString());
-}
-
-// Make sure that folder uris always end with a slash
-export function normalizeFolderName(folderName: string): string {
-  if (folderName.charAt(folderName.length - 1) !== '/') {
-    folderName = folderName + '/';
-  }
-  return folderName;
-}
-
 export function getWorkspaceFolder(folderName: string): vscode.WorkspaceFolder | undefined {
   return vscode.workspace.getWorkspaceFolder(vscode.Uri.parse(folderName));
 }
@@ -51,16 +39,4 @@ export function isTerraformFile(document?: vscode.TextDocument): boolean {
 
   // be safe and default to false
   return false;
-}
-
-export function sortedWorkspaceFolders(): string[] {
-  const workspaceFolders = vscode.workspace.workspaceFolders;
-  if (workspaceFolders) {
-    return workspaceFolders
-      .map((f) => getFolderName(f))
-      .sort((a, b) => {
-        return a.length - b.length;
-      });
-  }
-  return [];
 }
